+++ src/app/dev/DevToys.Blazor/Components/AutoSuggestBox/AutoSuggestBox.razor
@@ -32,7 +32,7 @@
                          Context="item"
                          RaiseSelectionEventOnKeyboardNavigation="false"
                          UseNativeScrollBar="true"
-                         OnSelectedIndexChanged="OnSelectedIndexChanged">
+                         SelectedIndexChanged="OnSelectedIndexChanged">
                     <ItemTemplate>
                         @ItemTemplate(item)
                     </ItemTemplate>
