+++ src/app/dev/DevToys.Blazor/Components/ComboBox/ComboBox.razor.scss
@@ -1,0 +1,40 @@
+@use "../../Assets/sass/devtoys" as *;
+
+.combo-box {
+    @include typography-body;
+
+    &-with-header {
+        display: inline-flex;
+        flex-direction: column;
+        position: relative;
+
+        .combo-box-header {
+            margin-bottom: 4px;
+        }
+    }
+
+    &-button {
+        .arrow-down-icon {
+            height: 16px;
+            width: 16px;
+            padding-top: 4px;
+            margin-left: 4px;
+            transition-duration: 0.2s;
+            transition-property: transform;
+            color: currentColor !important;
+        }
+
+        &:active {
+            .arrow-down-icon {
+                transform: translateY(2px);
+                color: currentColor !important;
+            }
+        }
+    }
+
+    &-list-box {
+        margin: 0px;
+        padding: 0px;
+        max-height: 400px;
+    }
+}
