+++ src/app/dev/DevToys.Blazor/Components/ContextMenu/ContextMenu.razor
@@ -32,7 +32,7 @@
                          OverrideDefaultItemTemplate="true"
                          Virtualize="false"
                          RaiseSelectionEventOnKeyboardNavigation="false"
-                         OnSelectedIndexChanged="OnContextMenuItemSelected"
+                         SelectedIndexChanged="OnContextMenuItemSelected"
                          UseNativeScrollBar="true"
                          Class="context-menu-list-box"
                          @onkeydown="OnListBoxKeyDown">
