+++ src/app/dev/DevToys.Blazor/Components/DropDownList/DropDownList.razor
@@ -38,7 +38,7 @@
                                  OverrideDefaultItemTemplate="true"
                                  Virtualize="true"
                                  RaiseSelectionEventOnKeyboardNavigation="false"
-                                 OnSelectedIndexChanged="OnDropDowlListItemSelected"
+                                 SelectedIndexChanged="OnDropDowlListItemSelected"
                                  UseNativeScrollBar="true"
                                  Class="context-menu-list-box"
                                  Style="max-height: 400px">
