+++ src/app/dev/DevToys.Blazor/Components/ListBox/ListBox.razor
@@ -1,6 +1,6 @@
 @namespace DevToys.Blazor.Components
-@inherits JSStyledComponentBase
 @typeparam TElement
+@inherits SelectBase<TElement>
 
 <ScrollViewer Orientation="Orientation.Vertical"
               IsEnabled="@IsEnabled"
